feat(frontend): fixed guides comments
