+++ src/theme/Footer/index.js
@@ -7,44 +7,16 @@
 
  import React, { useEffect, useState } from 'react'
  import clsx from 'clsx';
- 
- //import Link from '@docusaurus/Link';
- import { useThemeConfig, useColorMode } from '@docusaurus/theme-common';
+  import { useThemeConfig, useColorMode } from '@docusaurus/theme-common';
  import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
  import useBaseUrl from '@docusaurus/useBaseUrl';
  import styles from './styles.module.css';
  import whiteLogo from '../../../static/img/eoscostarica-logo.png'
  import blackLogo from '../../../static/img/eos-logo-dark.png'
- import { Box, Grid, Container, Link, Image } from '@material-ui/core';
+ import { Box, Grid, Container, Link} from '@material-ui/core';
  import { AiFillGithub, AiFillLinkedin, AiOutlineTwitter, AiOutlineInstagram, AiFillYoutube } from "react-icons/ai";
  import { BsMedium } from "react-icons/bs";
- 
- 
- 
- 
- 
- function FooterLink({ to, href, label, prependBaseUrlToHref, ...props }) {
-   const toUrl = useBaseUrl(to);
-   const normalizedHref = useBaseUrl(href, { forcePrependBaseUrl: true });
- 
-   return (
-     <Link
-       className="footer__link-item"
-       {...(href
-         ? {
-           target: '_blank',
-           rel: 'noopener noreferrer',
-           href: prependBaseUrlToHref ? normalizedHref : href,
-         }
-         : {
-           to: toUrl,
-         })}
-       {...props}>
-       {label}
-     </Link>
-   );
- }
- 
+  
  const FooterLogo = ({ url, alt }) => (
    <img className="footer__logo" alt={alt} src={url} width='140' height='100' loading="lazy" />
  );
@@ -55,13 +27,10 @@
    const { siteConfig = {} } = context;
    const { themeConfig = {} } = siteConfig;
    const { footer } = themeConfig;
- 
-   const { copyright, links = [], logo = {} } = footer || {};
+   const { copyright, logo = {} } = footer || {};
    const [logoUrl, setLogoUrl] = useState(useBaseUrl(logo.src));
    const { colorMode } = useColorMode();
- 
- 
- 
+  
    if (!footer) {
      return null;
    }
@@ -85,29 +54,21 @@
          'styles.footerDark': footer.style === 'dark'
        })}
      >
- 
-       <Container maxWidth="xl">
- 
-         <Grid container spacing={2}>
- 
- 
-           <Grid container >
- 
-             <Grid item xs={12} sm={3}>
+        <Container maxWidth="xl">
+          <Grid container spacing={2}>
+            <Grid container >
+              <Grid item xs={12} sm={3}>
                <Box align='center' >
                  <a href={logo.href}>
                    <FooterLogo alt={logo.alt} url={logoUrl} />
                  </a>
                  {copyright ? <div>Copyright &reg; {new Date().getFullYear()} EOS Costa Rica</div> : null}
- 
-               </Box>
+                </Box>
              </Grid>
- 
-             <Grid item xs={12} sm={6}>
+              <Grid item xs={12} sm={6}>
                <Box border={1} borderTop={0} borderBottom={0} borderColor={'#BDBDBD'} paddingLeft={20}>
                  <Grid container >
- 
-                   <Grid item xs={12} sm={4} >
+                    <Grid item xs={12} sm={4} >
                      <Box fontSize={18} >
                        <Box fontWeight={800} >EOS Costa Rica</Box>
                        <Box>
@@ -167,12 +128,10 @@
                        </Box>
                      </Box>
                    </Grid>
- 
-                 </Grid>
+                  </Grid>
                </Box>
              </Grid>
- 
-             <Grid item xs={12} sm={3}>
+              <Grid item xs={12} sm={3}>
                <div className={styles.contentSocial}>
                  <Grid container spacing={1} >
                    <Grid item xs={12} sm={4} align='right'>
