+++ src/theme/Footer/styles.module.css
@@ -25,3 +25,8 @@
 .contentSocial{
   margin-top: 25px;
 }
+
+.test{
+  border-top: 10px solid blue;
+
+}
